expose instructions

+++ programs/oracle/src/utils/constants.rs
@@ -10,6 +10,6 @@ pub const VOTING_LENGTH_SECONDS: i64 = SECONDS_PER_DAY as i64;
 pub const VOTING_FRACTION_REQUIRED: u64 = 10;
 
 // vote stake calculation as 1/x
-pub const VOTE_STAKE_RATE: f64 = 100.; // 1%
+pub const VOTE_STAKE_RATE: u32 = 100; // 1%
 
 pub const TEST_START_TIME: i64 = 1660681219;
